:truck: module_store 规范包名，png 装换为 webp
