+++ module_mine/src/main/java/com/mredrock/cyxbs/mine/page/sign/DailyViewModel.kt
@@ -11,7 +11,7 @@ import com.mredrock.cyxbs.common.viewmodel.event.SingleLiveEvent
 import com.mredrock.cyxbs.mine.network.model.ScoreStatus
 import com.mredrock.cyxbs.mine.util.apiService
 import com.mredrock.cyxbs.mine.util.extension.normalWrapper
-import com.ndhzs.api.store.IStoreService
+import com.mredrock.cyxbs.api.store.IStoreService
 import io.reactivex.rxjava3.functions.Function
 
 /**
