+++ module_qa/src/main/java/com/mredrock/cyxbs/qa/ui/widget/likeview/LikeViewSlim.kt
@@ -12,7 +12,7 @@ import com.mredrock.cyxbs.common.utils.extensions.*
 import com.mredrock.cyxbs.qa.R
 import com.mredrock.cyxbs.qa.config.CommentConfig
 import com.mredrock.cyxbs.qa.network.ApiServiceNew
-import com.ndhzs.api.store.IStoreService
+import com.mredrock.cyxbs.api.store.IStoreService
 import java.lang.ref.WeakReference
 
 /**
