+++ module_store/api_store/src/main/java/com/mredrock/cyxbs/api/store/RouteTable.kt
@@ -1,0 +1,10 @@
+package com.mredrock.cyxbs.api.store
+
+/**
+ * ...
+ * @author 985892345 (Guo Xiangrui)
+ * @email guo985892345@foxmail.com
+ * @date 2022/8/6 17:03
+ */
+
+const val STORE_SERVICE = "/store/service"
