--- module_store/api_store/src/main/java/com/ndhzs/api/store/RouteTable.kt
@@ -1,10 +1,0 @@
-package com.ndhzs.api.store
-
-/**
- * ...
- * @author 985892345 (Guo Xiangrui)
- * @email guo985892345@foxmail.com
- * @date 2022/8/6 17:03
- */
-
-const val STORE_SERVICE = "/store/service"
