+++ module_store/src/main/debug/store/ui/DebugActivity.kt
@@ -14,6 +14,10 @@ import com.mredrock.cyxbs.store.page.center.ui.activity.StoreCenterActivity
  */
 class DebugActivity : BaseDebugActivity() {
   
+  override val isNeedLogin: Boolean
+//    get() = false
+    get() = true
+  
   override fun onDebugCreate(savedInstanceState: Bundle?) {
     startActivity(
       Intent(this, StoreCenterActivity::class.java)
