+++ module_volunteer/src/main/java/com/mredrock/cyxbs/volunteer/viewmodel/DiscoverVolunteerFeedViewModel.kt
@@ -10,7 +10,7 @@ import com.mredrock.cyxbs.common.viewmodel.BaseViewModel
 import com.mredrock.cyxbs.common.viewmodel.event.SingleLiveEvent
 import com.mredrock.cyxbs.volunteer.bean.VolunteerTime
 import com.mredrock.cyxbs.volunteer.network.ApiService
-import com.ndhzs.api.store.IStoreService
+import com.mredrock.cyxbs.api.store.IStoreService
 
 class DiscoverVolunteerFeedViewModel : BaseViewModel() {
     val volunteerData = MutableLiveData<VolunteerTime?>()
